pyproject.toml
[project]
name = "law-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "pandas>=3.0.0",
    "pdfplumber>=0.11.9",
    "polars>=1.37.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "torch>=2.10.0",
    "transformers>=4.57.6",
]
